chore: rename Enum -> enums, Exception -> exceptions, and adjust formatting in Student entity
