--- backend/src/main/java/com/tkpm/sms/Enum/Faculty.java
@@ -1,8 +1,0 @@
-package com.tkpm.sms.Enum;
-
-public enum Faculty {
-    Faculty_of_Japanese,
-    Faculty_of_French,
-    Faculty_of_Law,
-    Faculty_of_Business_English
-}
