+++ backend/src/main/java/com/tkpm/sms/dto/StudentResponse.java
@@ -1,9 +1,9 @@
 package com.tkpm.sms.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.tkpm.sms.Enum.Faculty;
-import com.tkpm.sms.Enum.Gender;
-import com.tkpm.sms.Enum.Status;
+import com.tkpm.sms.enums.Faculty;
+import com.tkpm.sms.enums.Gender;
+import com.tkpm.sms.enums.Status;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
