+++ backend/src/main/java/com/tkpm/sms/entity/Student.java
@@ -1,8 +1,8 @@
 package com.tkpm.sms.entity;
 
-import com.tkpm.sms.Enum.Faculty;
-import com.tkpm.sms.Enum.Gender;
-import com.tkpm.sms.Enum.Status;
+import com.tkpm.sms.enums.Faculty;
+import com.tkpm.sms.enums.Gender;
+import com.tkpm.sms.enums.Status;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
@@ -24,27 +24,38 @@ public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     String id;
-    @Column(name = "studentId", unique = true)
+
+    @Column(name = "student_Id", unique = true)
     String studentId;
+
     String name;
+
     Date dob;
+
     @Enumerated(EnumType.STRING)
     @Column (nullable = false)
     Gender gender;
+
     @Enumerated(EnumType.STRING)
     @Column (nullable = false)
     Faculty faculty;
+
     Integer course;
+
     String program;
+
     @Email
     @NotNull
     @Column(nullable = false, unique = true)
     String email;
+
     String address;
+
     @NotNull
     @Pattern(regexp = "^0\\d{9}$", message = "Phone number must start with 0 and have 10 digits")
     @Column(nullable = false, unique = true, length = 10)
     String phone;
+
     @Enumerated(EnumType.STRING)
     @Column (nullable = false)
     Status status = Status.Studying;
