+++ backend/src/main/java/com/tkpm/sms/enums/Faculty.java
@@ -1,0 +1,13 @@
+package com.tkpm.sms.enums;
+
+public enum Faculty {
+    Faculty_of_Japanese,
+    Faculty_of_French,
+    Faculty_of_Law,
+    Faculty_of_Business_English;
+
+    @Override
+    public String toString() {
+        return name().replace("_", " ");
+    }
+}
