+++ backend/src/main/java/com/tkpm/sms/enums/Gender.java
@@ -1,4 +1,4 @@
-package com.tkpm.sms.Enum;
+package com.tkpm.sms.enums;
 
 public enum Gender {
     MALE, FEMALE
