+++ backend/src/main/java/com/tkpm/sms/enums/Status.java
@@ -1,4 +1,4 @@
-package com.tkpm.sms.Enum;
+package com.tkpm.sms.enums;
 
 public enum Status {
     Studying, Graduated, Dropped, Suspended
